Mocks now can be implemented automatically
Now mocks trigger `DynamicCallable` and get result dynamically. It has the lowest priority among other ways to get result - directly stubbed calls and calls to default implementation.

One implementation of `DynamicCallable` is `AnyGeneratorDynamicCallable` that uses `AnyGenerator` and returns value if the following conditions are met:

- `AnyGenerator` didn't threw error (e.g. all specific generators were registered in it's implementation, which is currently `AnyGeneratorImpl`)
- None of functions arguments are escaping closures. This is important, because usually asynchronous functions have `completion` argument and are meaningless if `completion` is not called. Note that the feature of stubbing asynchronous functions is planned for future releases.

Because `AnyGenerator` relies on DI of testcase, `UnconfiguredMockManagerFactory` uses `CanNotProvideResultDynamicCallable`, and it basically means that this feature is disabled unless mock isn't connected with DI (via `setMockManager` or any code that use this function).

Tests were added and there were a lot of issues with them, so a lot of additional changes were made.

Additional changes:

- Add ability to inspect argument and return types of closures in `RecordedCall`
- Fix code generation of builders for generic functions, before this commit code generation could generate code that can't be compiled
- Add `@discardableResult` attribute to `setMockManager`
- Add `matches`, global function, a shortcut for creating `FunctionalMatcher`.
- Add syntactic sugar to `MockRegisterer` - a register function that also returns argument. This syntax is now possible: `let mock = registerer.register(MockSomething())`
- Add ability to disambiguate types when using `any` and `none` matchers (via `type` argument)
- Add ability to disambiguate return values in stubbing. Example:
```
mock
    .verify()
    .genericFunctionWithDifferentReturnValue(generic: any(type: String.self))
    .returning(Int.self)
    .isCalledOnlyOnce(timeout: 0)
```

Additional changes related to tests:

- A lot of new tests were added
- `GeneratorTestsDependencies` class was removed and it's code was merged into `TestsCaseDependencies`
- `MixboxGeneratorIntegrationTestsFixtureObjcProtocol` was renamed to `MocksTestsFixtureObjcProtocol`, same for similar protocols.
- `BaseMixboxMocksRuntimeTests` was added to store common logic
- I got false positive results of tests due to bug in recording failures - I called `super.record` method of XCTestCase, but it was calling `self.recordFailure` in its implementation, and led to stack overflow. I've rewritten code from Swift to Objective-C, because it is the only way. The new code is `TestCaseSuppressingWarningAboutDeprecatedRecordFailure` to call `super` and `XCTestCase+SuppressingWarningAboutDeprecatedRecordFailure` to call self while also omitting warning about deprecated API (see the comments where I explained why it is unavoidable).
- Add better way of suppressing deprecation warning about `recordFailure`, because previous way led to false-positive test results
- `assertFails` can produce false-positive results. Some errors are expected and desired (errors produced by testing utilities that are being tested), while others are not, however, before this PR `assertFails` didn't differentiate between expected and unexpected errors, so I made `UninterceptableErrorTracker` utility. It can bypass interception of errors by `assertFails`, tracks errors and fails test in `tearDown`.
- `HeapFileLine` was removed, because it was duplicating `RuntimeFileLine`
- Mess with types of `line` (as a representation of source code location) was reduced. Now it is `UInt` most of the time (because most of the code uses UInt, I'm talking about thiird-party code such as XCTest)
- `SourceCodeLineTypeConverter` was added to store logic for converting types of variables representing line number

+++ Frameworks/TestsFoundation/Reporting/FileLineForFailureProvider/StackTrace/ExtendedStackTraceEntry.swift
@@ -1,15 +1,14 @@
 public final class ExtendedStackTraceEntry {
     public let file: String?
-    public let line: UInt64?
+    public let line: UInt?
     public let owner: String?
     public let symbol: String?
     public let demangledSymbol: String?
     public let address: UInt64
 
-// sourcery:inline:auto:ExtendedStackTraceEntry.Init
     public init(
         file: String?,
-        line: UInt64?,
+        line: UInt?,
         owner: String?,
         symbol: String?,
         demangledSymbol: String?,
@@ -22,5 +21,4 @@ public final class ExtendedStackTraceEntry {
         self.demangledSymbol = demangledSymbol
         self.address = address
     }
-// sourcery:end
 }
